Update 5: 17/6/13
The GP Parser is working, but it is very basic. It parses a single file and has minimal error
reporting. At the moment it doesn't function correctly as it reports a syntax error on any
input file. Need to add better error reporting to figure out the problem.

- Added the single character tokens '@', '[' and ']'.
- Removed the token processing code from the lexer

Parser conflicts resolved:
- Changed production in MacroDecl: ComSeq ProcList -> ProcList ComSeq
- Added '@' before MacroID in MacroCall to distinguish from RuleID
- Changed RuleSetCall to '[' IDList ']' as curly braces clashed with
  local macro declarations.