Update 23: Added semantic checking to declaration_scan, debugged pretty.c
Extended declaration_scan to check for multiple occurrences of a single
procedure name and for multiple occurrences of the same rule name in the
same scope. This is done while names are added to the symbol table.

The AST pretty printer had a rather large bug in that pretty_print_list
wasn't called to print, for example, parameter lists in rules. This meant
that an empty parameter list was handled incorrectly. To rectify this, I
rewrote the pretty_print_list macro so it can be called in any printing
function, not just print_list, and invoked it in the appropriate places:
- print_procedure for local_decls.
- print_rule for variables, interface and condition (condition handled
  explicitly as opposed to a macro call as print_condition is called
  instead of print_list).
- print_graph for nodes and edges.

seman.c: Added local string variables proc_name and rule_name to store
         the names of procedures instead of repeating long structure
         accesses.
makefile: Added functionality to call make with an extensionless
          filename which executes gpparse and dot -Tjpeg on the file
          automatically.