Update 32: Memory Management
I added some code to garbage collect the various bits of heap memory
allocate my code. This includes the AST, the symbol table, symbols
themselves and identifiers encounters in the parsing stage.

I wrote a series of functions to free the AST with a depth-first walk.
I refined the symbol table code to handle its memory properly.
This involved changing the call of g_hash_table_new to
g_hash_table_new_full and writing a function to free memory consumed
by a symbol list (the values of the symbol table). I added some 'free's
to the action code of the parser so that identifiers and strings, heap
allocated by the lexer, are freed immediately after use. Finally,
I added a bunch of strdups in the AST constructor functions and in
seman.c to ensure that pointers to heap memory weren't cloned and freed
erroneously.

seman.c
-------
- Reorganised code such that reporting of non-simple expressions is
  performed entirely in gp_list_scan, controlled by three file-scope
  variables declared before gp_list_scan's definition.
- Removed some redundant code from case VARIABLE of atomic_exp_scan.

pretty.c/h
----------
- Moved macro definitions from pretty.c to pretty.h.

gpparser.y
----------
- Added some more productions to catch the use of ProcID instead of RuleID
  and vice versa.