Update 33: Host Graph Parser
I have implemented the host graph parser. Now the program takes
two filenames as arguments. It expects the GP program file first
followed by the host graph file.

This host graph parsing is achieved with an extension to the grammar
in gpparser.y. I added two rules at the top level of the grammar:

Initialise: GP_PROGRAM Program
          | GP_GRAPH HostGraph

Where Program and HostGraph are the start symbols of the GP Program
grammar and the host graph grammar respectively. The host graph grammar
is similar to the grammar for graphs, except that all atoms must be
constant values (EMPTY, NUM or STR). This necessitated the introduction
of new nonterminals CNodeList, CNode, CLabel etc. (c for constant).

The tokens GP_PROGRAM and GP_GRAPH are returned by the lexer before
the input file is parsed depending on which file is being read. This
is controlled by main.c, which assigns a global variable to 1 or 2
depending on whether the program is about to commence parsing of the
GP program or parsing of the host graph. This variable is read by
the lexer, which returns the appropriate token.

main.c
- Program now expects gpparse to be called with two filenames. There is no
  more -d flag (see below).
- New macro definitions PARSER_TRACE and DRAW_HOST_GRAPH_AST to control
  yydebug and printing of host graph AST. Other macros were renamed.
- Changed abort_scan from int to bool (this also affected gpparser.y,
  gplexer.l and seman.c).
- The name of the log file is now [program_name].log rather than gp.log.

pretty.c/h
- New function print_dot_host_graph to create the file and call
  print_graph on the pointer to the host graph's AST.

ast.c
- Added guards to every call to freeing functions, as the pointers could
  be NULL.

seman.c/h
- Added a function to reverse lists in the host graph's AST.
- Changed error reporting: any print to stderr also gets printed
  to log_file.
- Merged cases ADD, SUBTRACT, MULTIPLY and DIVIDE in atom_exp_scan.
- Added code to check for multiple instances of the same type being
  declared in a rule's variable declaration section.

gplexer.lex
- Imposed a size limit of 63 on identifiers and strings.