Update 34: New type 'character'
A new character type has been added to the language, resulting
in a new GPAtomExp (character constant) and a new GPCondition
(character check, in the subtype checker category). This replaces
the short-lived head and tail operators: they have been removed
from the language, as has all the code for processing them.

I have removed the value destroy function for the GHashTable as this
was causing issues when inserting Symbol lists into the table.
When freeing the hash table I now call g_hash_table_foreach with
free_symbol_list, for the values; and free, for the keys (which are just
malloc'd strings). In addition, keys are explicitly freed after
hash table insertions in seman.c. They are also freed if the hash
table insertion doesn't take place

I have also added some guards to function calls or structure acceses
in seman.c.

makefile
- Rewrote the default make rule so that it takes two arguments and
  additionally calls dot -Tjpeg on the dot file for the graph AST.