Update 40: Debugging.
I modified the List syntax so that Bison can perform error recovery
from an atomic expression. Now lists are required to end in a ':'
so that the "error ':'" rule acts properly even if a syntax error
is encountered in the final element of a list. This necessitated
a change to the List productions in the grammar.

gpparser.y
- Changed names of host graph symbols e.g. CExp -> HostExp

gplexer.lex
- Added rules to handle new lines within string and characters.
  If this happens then the parsing terminates.
- Added yyterminate() calls after every regexp matching <<EOF>>.