Update 49: Debugging the graph module.
The integer hash tables for the inedges/outedges of nodes were not
functioning correctly. This is because I was using g_int_hash instead
of g_direct_hash. The latter is the appropriate hash function for
my purposes.

The relabelling functions were also incomplete: I added code to modify
the appropriate index-by-label-class hash tables in the case that the
new label is a different label class to the old one.