Update 54: Debugging
If there are multiple variables in a label, and the matching fails
after the first one has been assigned, the assignment of the first
variable is erroneously kept. I have implemented the functionality
to restore the assignment to its original state if labelMatch
fails at any point.

Also, in staticSearchplan there were a few points where I incorrectly
used break instead of continue.