Label Matching
The code generation has been reworked so that the values of GP2
variables are assigned to C variables before the host graph is
modified for rule application. In this way, variables occurring
more than once in the RHS do not require the assignment to be
queried multiple times.

Generation of integer and string expressions fixed so that the
runtime variables are guaranteed to be unique.

Detailed comments, including examples, added to the label code
generation module.

Rule application code updated: no need for an explicit runtime
array to obtain images of LHS-nodes: this is always present in
the morphism.

Semantic checking updated to reflect new context conditions,
namely that node IDs in degree operators and edge predicates must
occur in the interface (previously the LHS or RHS).

Fixed a bug in variableScan: it was only updating the types of
variables in the LHS graph.

Added code to populate a rule's variable list during the
transformation phase.