Degree Operator and Variable Evaluation
The transformation module has been significantly refactored.
There is also new code to record the variables that are used in the
RHS and to record nodes whose degrees are queried in RHS labels.
These are used to generate the necessary variables in runtime
without generating variables that will not be needed.

The code for generating integer and string expressions has been
greatly simplified. I observed that an integer expression can be
generated in one pass without having to create lots of runtime
variables. The string case is not as clean, two AST passes are
still required, but there are no more runtime variables for RHS
strings.