Reimplemented Label Class Tables.
The graph's label class indexing structure is a linked list of
LabelClassTables instead of the old two-dimensional array. The tables
explicitly store the mark and the label class they represent.
Very few of the label class/mark combinations are used in a typical
GP 2 program, so it did not make sense to manage slots for all of
them at once.

Minor changes:
- Fixed a bug in generateRHSLabels that generated code to make a
  label based on the false assumption that there is at most one
  list variable in a RHS label.
- Fixed a bug in makeRule causing some GP2 variables to be marked
  incorrectly as unused.