Resetting 'matched' flag of nodes and edges moved to initialiseMorphism function
The initialiseMorphism function takes a second optional graph argument. If a graph is
passed, all nodes and edges referred to by the morphism have their 'matched' flag
set to false. This simplifies the code generator since it does not need to print
explicit calls to 'resetMatchedNode/EdgeFlag' in the rule application function.

Related to this issue, a bug has been fixed in which the matched flags weren't
reset after matching a rule R in a program of the form "if R then R", causing
the second rule application to erroneously fail.