Bug fixes
Fixed improper variable use in undoChanges function (case REMOVED_EDGE):
'index' renamed to 'edge.index' in six assignments.

Fixed a bug concerning the mishandling of the graph changes stack when
executing a loop within a loop. The inner loop may discard changes that
are required to restore the graph if the outer loop fails. The data
structure for loops has an extra flag 'inner_loop' set to true in the
program analysis if a loop occurs inside another loop. Supported by
an extra argument to the annotate function in analysis.c. Program
code generator updated to print code that updates the restore point
after a successful iteration of an inner loop. Outer loops retain the
old behaviour of discarding changes.

Updated the tracing codeto print more detailed information on backtracking.